LayoutTests/fast/forms/button-multiline-height-expected.txt

Test that the height of a button that contains multiple lines of text is greater than the height of a button that contains a single line of text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS multipleLinesBtn.offsetHeight > singleLineBtn.offsetHeight is true
PASS successfullyParsed is true

TEST COMPLETE
Line 1

Line 1
Line 2
